ADMIN chart fixed
